int to double

+++ src/rcpp_first_claim.cpp
@@ -10,7 +10,7 @@ NumericVector rcpp_first_claim(std::vector<double> id, std::vector<double> claim
   // locations vector
   std::vector<double> rows(1, 0);
   std::vector<double>::iterator ip;
-  double row_p = 0;
+  int row_p = 0;
   for (ip = id.begin(); ip != id.end(); ++ip) {
     if (*ip != *(ip+1)) {
       row_p += 1;
@@ -24,8 +24,8 @@ NumericVector rcpp_first_claim(std::vector<double> id, std::vector<double> claim
   result.resize(claim.size());
   std::vector<double>::iterator iq;
   for (iq = rows.begin(); iq != rows.end(); ++iq) {
-    double n1 = 0;
-    for (double i = *iq; i < *(iq+1); ++i) {
+    int n1 = 0;
+    for (int i = *iq; i < *(iq+1); ++i) {
       if (n1 == 0 && claim[i] == 1) {
         result[i] = 1;
         n1 = 1;
